use controls from transport for play stop
