+++ src/components/PianoRoll/PianoRoll.tsx
@@ -1,11 +1,9 @@
-import {useState, useEffect} from "react";
-import * as Tone from "tone";
 import Key from "./Key";
 
 
 import "./piano-roll.css";
 import {useDispatch, useSelector} from "react-redux";
-import {setPlay, setSequence} from "../../state/actions";
+import {setSequence} from "../../state/actions";
 
 
 type Column = {
@@ -15,12 +13,7 @@ type Column = {
 
 const PianoRoll = () => {
     const sequence = useSelector(state => state.sequence.sequence);
-    const isPlaying = useSelector(state => state.player.isPlaying);
     const dispatch = useDispatch()
-
-
-    const [currentColumn, setCurrentColumn] = useState(null);
-    const synth = new Tone.PolySynth().toDestination();
 
     function handleNoteClicked(clickedColumn: number, clickedNote: number) {
         try {
@@ -44,49 +37,8 @@ const PianoRoll = () => {
 
     }
 
-    const playClicked = async () => {
-        let music: string[][] = [];
-
-        sequence.forEach((column) => {
-            let columnNotes: string[] = [];
-            column.map(
-                (shouldPlay) => shouldPlay.isActive && columnNotes.push(shouldPlay.note)
-            );
-            music.push(columnNotes);
-        });
-
-        await Tone.start();
-
-        const Sequencer = new Tone.Sequence(
-            (time, column) => {
-                setCurrentColumn(column);
-
-                synth.triggerAttackRelease(music[column], "8n", time);
-            },
-            [0, 1, 2, 3, 4, 5, 6, 7],
-            "8n"
-        );
-
-        if (isPlaying) {
-            dispatch(setPlay());
-            setCurrentColumn(null);
-
-            await Tone.Transport.stop();
-            await Sequencer.stop();
-            await Sequencer.clear();
-            await Sequencer.dispose();
-
-            return;
-        }
-        dispatch(setPlay());
-
-        await Sequencer.start();
-        await Tone.Transport.start();
-    };
-
     return (
         <>
-            <button onClick={playClicked}>{!isPlaying ? "Play" : "Stop"}</button>
             <div className="row">
                 {sequence.map((column, columnIndex) => {
                     return (
